[breaking] Rename telemetry to metrics
