+++ docs/configuration.md
@@ -18,9 +18,9 @@
   - `format` - output format for the logs. Allowed values are `text` or `json`.
   - `level` - messages with this level and above will be logged. Valid levels are: `trace`, `debug`, `info`, `warn`,
     `error`, `fatal`, `panic`.
-- `telemetry` - settings related to the collection of data used for continued improvement of Arduino CLI.
-  - `addr` - TCP port used for telemetry communication.
-  - `enabled` - controls the use of telemetry.
+- `metrics` - settings related to the collection of data used for continued improvement of Arduino CLI.
+  - `addr` - TCP port used for metrics communication.
+  - `enabled` - controls the use of metrics.
 
 ## Configuration methods
